requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.16.1
    # via mlflow
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
    #   streamlit
blinker==1.9.0
    # via
    #   flask
    #   streamlit
cachetools==5.5.2
    # via
    #   google-auth
    #   mlflow-skinny
    #   streamlit
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   dask
    #   distributed
    #   flask
    #   mlflow-skinny
    #   prefect
    #   streamlit
    #   uvicorn
cloudpickle==3.1.1
    # via
    #   dask
    #   distributed
    #   mlflow-skinny
    #   prefect
contourpy==1.3.2
    # via matplotlib
croniter==6.0.0
    # via prefect
cycler==0.12.1
    # via matplotlib
dask==2025.5.1
    # via
    #   distributed
    #   prefect
databricks-sdk==0.57.0
    # via mlflow-skinny
distributed==2025.5.1
    # via prefect
docker==7.1.0
    # via
    #   mlflow
    #   prefect
fastapi==0.115.12
    # via
    #   -r requirements.in
    #   mlflow-skinny
flask==3.1.1
    # via mlflow
fonttools==4.58.2
    # via matplotlib
fsspec==2025.5.1
    # via dask
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via
    #   mlflow-skinny
    #   streamlit
google-api-core==2.25.1
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.40.3
    # via
    #   databricks-sdk
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.4.3
    # via google-cloud-storage
google-cloud-storage==3.1.0
    # via -r requirements.in
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.70.0
    # via google-api-core
graphene==3.4.3
    # via mlflow
graphql-core==3.2.6
    # via
    #   graphene
    #   graphql-relay
graphql-relay==3.2.0
    # via graphene
gunicorn==23.0.0
    # via mlflow
h11==0.16.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
importlib-metadata==8.7.0
    # via
    #   dask
    #   mlflow-skinny
    #   opentelemetry-api
    #   streamlit
importlib-resources==6.5.2
    # via prefect
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   altair
    #   distributed
    #   flask
    #   pydeck
joblib==1.5.1
    # via scikit-learn
jsonschema==4.24.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
kiwisolver==1.4.8
    # via matplotlib
locket==1.0.0
    # via
    #   distributed
    #   partd
loguru==0.7.3
    # via -r requirements.in
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   mako
    #   werkzeug
marshmallow==4.0.0
    # via
    #   marshmallow-oneofschema
    #   prefect
marshmallow-oneofschema==3.2.0
    # via prefect
matplotlib==3.10.3
    # via
    #   -r requirements.in
    #   mlflow
    #   seaborn
mlflow==3.1.0
    # via -r requirements.in
mlflow-skinny==3.1.0
    # via mlflow
msgpack==1.1.1
    # via
    #   distributed
    #   prefect
mypy-extensions==1.1.0
    # via prefect
narwhals==1.43.0
    # via
    #   altair
    #   plotly
numpy==2.3.0
    # via
    #   contourpy
    #   matplotlib
    #   mlflow
    #   pandas
    #   pydeck
    #   scikit-learn
    #   scipy
    #   seaborn
    #   streamlit
opentelemetry-api==1.34.1
    # via
    #   mlflow-skinny
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-sdk==1.34.1
    # via mlflow-skinny
opentelemetry-semantic-conventions==0.55b1
    # via opentelemetry-sdk
packaging==25.0
    # via
    #   altair
    #   dask
    #   distributed
    #   gunicorn
    #   matplotlib
    #   mlflow-skinny
    #   plotly
    #   prefect
    #   streamlit
pandas==2.3.0
    # via
    #   -r requirements.in
    #   mlflow
    #   seaborn
    #   streamlit
partd==1.4.2
    # via dask
pendulum==3.1.0
    # via prefect
pillow==11.2.1
    # via
    #   matplotlib
    #   streamlit
plotly==6.1.2
    # via -r requirements.in
prefect==1.4.1
    # via -r requirements.in
proto-plus==1.26.1
    # via google-api-core
protobuf==6.31.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   mlflow-skinny
    #   proto-plus
    #   streamlit
psutil==7.0.0
    # via distributed
pyarrow==20.0.0
    # via
    #   mlflow
    #   streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.6
    # via
    #   fastapi
    #   mlflow-skinny
pydantic-core==2.33.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
pympler==1.1
    # via streamlit
pyparsing==3.2.3
    # via matplotlib
python-box==7.3.2
    # via prefect
python-dateutil==2.9.0.post0
    # via
    #   croniter
    #   graphene
    #   matplotlib
    #   pandas
    #   pendulum
    #   prefect
    #   streamlit
python-slugify==8.0.4
    # via prefect
pytz==2025.2
    # via
    #   croniter
    #   pandas
    #   prefect
pyyaml==6.0.2
    # via
    #   dask
    #   distributed
    #   mlflow-skinny
    #   prefect
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   databricks-sdk
    #   docker
    #   google-api-core
    #   google-cloud-storage
    #   mlflow-skinny
    #   prefect
    #   streamlit
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
scikit-learn==1.7.0
    # via
    #   -r requirements.in
    #   mlflow
scipy==1.15.3
    # via
    #   mlflow
    #   scikit-learn
seaborn==0.13.2
    # via -r requirements.in
semver==3.0.4
    # via streamlit
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
sortedcontainers==2.4.0
    # via distributed
sqlalchemy==2.0.41
    # via
    #   alembic
    #   mlflow
sqlparse==0.5.3
    # via mlflow-skinny
starlette==0.46.2
    # via fastapi
streamlit==1.8.0
    # via -r requirements.in
tabulate==0.9.0
    # via prefect
tblib==3.1.0
    # via distributed
text-unidecode==1.3
    # via python-slugify
threadpoolctl==3.6.0
    # via scikit-learn
toml==0.10.2
    # via
    #   prefect
    #   streamlit
toolz==1.0.0
    # via
    #   dask
    #   distributed
    #   partd
tornado==6.5.1
    # via
    #   distributed
    #   streamlit
typing-extensions==4.14.0
    # via
    #   alembic
    #   altair
    #   anyio
    #   fastapi
    #   graphene
    #   mlflow-skinny
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via
    #   pandas
    #   pendulum
tzlocal==5.3.1
    # via streamlit
urllib3==2.4.0
    # via
    #   distributed
    #   docker
    #   prefect
    #   requests
uvicorn==0.34.3
    # via
    #   -r requirements.in
    #   mlflow-skinny
validators==0.35.0
    # via streamlit
werkzeug==3.1.3
    # via flask
zict==3.0.0
    # via distributed
zipp==3.23.0
    # via importlib-metadata
